fix: compile

+++ frb_rust/src/handler/handler.rs
@@ -68,7 +68,7 @@ pub trait Handler {
 
     fn dart_fn_invoke<Ret>(&self, dart_fn_and_args: Vec<DartAbi>) -> DartFnFuture<Ret>;
 
-    fn dart_fn_handle_output(&self, call_id: i64);
+    fn dart_fn_handle_output(&self, call_id: i32);
 }
 
 #[derive(Clone, Debug)]
